Fix matching regex in Play plugin.
The regex `<.*>` has eager/greedy matching - it will match as
many characters as possible. This means that if there are multiple
bound variables in a path, this will match all the characters
between them, removing it from the resulting path label.

This commit changes the regex to use a "reluctant" quantifier.